requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
notion-client
python-dotenv
